proxy.py: Implement timeout mechanism for client connections.
It seems some clients remain connected for days (for no apparent reason).
It appears the connection is indeed still open, but completely silent -
possibly broken client.
This may be a problem when limiting connections (--max-connections).
Anyway, we disconnect them after 10 minutes of inactivity.  The 10
minutes limit was chosen considering that the HaaS server disconnects
inactive sessions after 3 minutes, and so this limit is not configurable.